diseño de registro y response de home con login
