+++ src/app/auth/pages/login/login.component.html
@@ -1,7 +1,7 @@
 <body>
     
 
-    <div class="bg-login pt-4 pb-4 shadow text-center">
+    <div class="bg-login pt-3 pb-4 shadow text-center">
     
         <img  width="15%" src="../../../assets/Imagenes/Logos/user.png"/>
          <p class="fs-2 login">Login</p>
@@ -18,8 +18,8 @@
             </div>
     
             <div class="pt-3 ">
-                <button   class="btn btn-primary text-light px-4 mx-1">
-                    Register 
+                <button  (click)="registrarse()"  class="btn btn-primary text-light px-4 mx-1">
+                    Registrarse 
                 </button>
     
                 <button  class="btn btn-light px-4 mx-1">
@@ -32,12 +32,11 @@
          <div class="justify-content-center">
             <p class="social pt-3 text-light">O INGRESA CON TU CUENTA DE GOOGLE</p>
             
-            <button id="buttonDiv" class="btn">
-                <img src="https://img.icons8.com/fluency/48/000000/google-logo.png" class="" width="0%"/> Login
+            <button id="buttonDiv" class="btn">     
     
             </button> 
             
-            <div id="buttonDiv"></div> 
+            
          </div>
