+++ src/app/auth/pages/register/register.component.css
@@ -1,0 +1,22 @@
+.space{
+    padding-top: 100px;
+    padding-bottom: 10%;
+   
+}
+
+.bg-login{
+    background-image: url("../../../../assets/Imagenes/Fondos/giphy.gif");
+    border-radius: 10px;
+}
+
+.pass{
+    padding-right: 7px;
+}
+
+.login{   
+    color: white;
+}
+
+.social{
+    font-size: 15px;
+}
