+++ src/app/auth/pages/register/register.component.html
@@ -1,0 +1,45 @@
+<body>
+    <div class="bg-login pt-3 pb-4 shadow text-center">
+
+        <img  width="15%" src="../../../../assets/Imagenes/Logos/user.png"/>
+         <p class="fs-2 login">Registrate</p>
+    
+         <form action="">
+            <div class="pt-0">
+                <input type="text" placeholder="First Name" class="rounded shadow px-3 " style="border: none;">
+            </div>
+    
+            <div class="pt-2 justify-content-center">
+               <input type="text" placeholder="Last Name" class="rounded shadow px-3 " style="border: none;">  
+            </div>
+    
+    
+             <div class="pt-2 justify-content-center">
+                <input type="text" placeholder="Email" class="rounded shadow px-3 " style="border: none;">   
+             </div>
+    
+             <div class="pt-2 justify-content-center">
+               <input type="password" placeholder="Password" class="rounded shadow px-3 " style="border: none;">   
+             </div>
+    
+             <div class="pt-2 justify-content-center">
+                <input type="password" placeholder="Repeat Password" class="rounded shadow px-3 " style="border: none;">   
+              </div>
+    
+            <div class="pt-4 ">
+
+               <button  class="btn btn-light px-4 mx-1" (click)="back()">
+                  Back
+              </button>
+               
+               <button  class="btn btn-primary px-4 mx-1">
+                    Let go
+                </button>
+            </div>           
+            
+         </form>
+           
+        
+      
+    </div>
+</body>
